// packages/language/src/completion/metadata/timeline-events.generated.ts
export interface TimelineEventMetadata {
  name: string;
  description: string;
  category?: string;
}